create eof package
