+++ setup.cfg
@@ -107,7 +107,7 @@ packages =
     ethereum/prague
     ethereum/prague/utils
     ethereum/prague/vm
-    ethereum/prague/vm/eof1
+    ethereum/prague/vm/eof
     ethereum/prague/vm/instructions
     ethereum/prague/vm/precompiled_contracts
     ethereum/prague/vm/precompiled_contracts/bls12_381
